Probably shouldn't need this
Mainly was doing this to get garbage JS out of the content where people
have managed to shove it in, but core matches them fine with raw
content. The regex looks like it should handle raw content to me and
it’s less overhead when running  through all these fields.